test(publisher): move legacy-WM migration cases to their own spec
Review round on #2153. Follows the production boundary: the migration is
its own state machine in `legacy-wm-migration.ts`, so its crash-window,
refusal, and resume cases now live in `test/legacy-wm-migration.test.ts`
instead of deepening the catch-all draft-lifecycle spec. A change to the
durable marker vocabulary or a migration state now lands in one focused
file rather than requiring navigation of a multi-thousand-line spec.

The migration fixtures (`legacyMigrationUris`, `stampMigrationMarker`,
`migrationMarkerCompleted`, `recordingAllocator`) move with the cases they
serve. These remain publisher-level integration tests against a real
DKGPublisher + OxigraphStore — only their home changed; no assertion was
altered. `draft-lifecycle.test.ts` loses 575 lines.

Parity verified three ways rather than assumed, since a test move that
silently drops coverage is the failure mode that matters here:
- 100 tests across the two files, the same total as before the split
- restoring the merged total-row classification reddens the two
  orphan-event cases from their new home
- skipping the content copy on resume reddens the pre-copy crash case

Validation: build:runtime:packages; tsc --noEmit clean on publisher; full
publisher suite 126 files / 1,816 passed / 6 skipped.
